+++ Craftsman/Commands/ProcessCommand.cs
@@ -145,7 +145,7 @@
                 }
             }
             
-            if (args.Length == 2 && (args[0] == "add:entity" || args[0] == "add:entities"))
+            if (args.Length == 2 && (args[0] == "add:entity" || args[0] == "add:entities" || args[0] == "new:entity"))
             {
                 var filePath = args[1];
                 if (filePath == "-h" || filePath == "--help")
@@ -186,6 +186,29 @@
                     }
 
                     AddBusCommand.Run(filePath, rootDir, fileSystem);
+                }
+            }
+
+            if ((args[0] == "add:bff"))
+            {
+                var filePath = "";
+                if (args.Length >= 2)
+                    filePath = args[1];
+
+                if (filePath == "-h" || filePath == "--help")
+                    AddBffCommand.Help();
+                else
+                {
+                    CheckForLatestVersion();
+
+                    var rootDir = fileSystem.Directory.GetCurrentDirectory();
+                    if (myEnv == "Dev")
+                    {
+                        Console.WriteLine("Enter the root directory.");
+                        rootDir = Console.ReadLine();
+                    }
+
+                    AddBffCommand.Run(filePath, rootDir, fileSystem);
                 }
             }